backend/requirements.txt
blinker==1.8.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
fsspec==2025.3.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==2.1.5
packaging==25.0
python-dotenv==1.0.1
PyYAML==6.0.3
requests==2.32.4
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.2.3
Werkzeug==3.0.6
zipp==3.20.2
